removed inline functions and adjustments in eslint
