--- utils/fetcher.ts
@@ -1,5 +1,0 @@
-export const fetcher = async (url: string) => {
-  const res = await fetch(url)
-  const data = await res.json()
-  return data
-}
