+++ utils/index.ts
@@ -1,9 +1,7 @@
-export * from './filterInvoice'
-export * from './fetcher'
+export * from './helpers'
 export * from './formik/initialValues'
 export * from './formik/validate'
 export * from './formik/onSubmit'
-export * from './url'
 export * from './apiCall/invoices/getInvoice'
 export * from './apiCall/invoices/getInvoices'
 export * from './apiCall/invoices/deleteInvoice'
