--- utils/url.ts
@@ -1,4 +1,0 @@
-export const url =
-  process.env.NODE_ENV === 'production'
-    ? 'https://invoice-frontend-mentor.vercel.app'
-    : 'http://localhost:3000'
